guass and jorden error done
