+++ answers.py
@@ -93,10 +93,11 @@ class Ui_ans(object):
         self.solutionWidget.setGeometry(QtCore.QRect(19, 149, 761, 421))
         self.solutionWidget.setObjectName("solutionWidget")
         self.errorsWidget = QtWidgets.QWidget(ans)
+        self.errorsWidget.setGeometry(QtCore.QRect(250, 10, 600, 600))
         self.errorsWidget.move(250,250)
         self.errorMessage = QtWidgets.QLabel(self.errorsWidget)
         font = QtGui.QFont()
-        font.setFamily("MV Boli")
+        font.setFamily("Times New Roman")
         font.setPointSize(20)
         self.errorsWidget.setFont(font)
